requirements.txt
# Sistema Minimarket Don Manuelito - Sprint 1
# Dependencias mínimas necesarias

# Manejo de datos
pandas>=1.5.0

# Manejo de imágenes
Pillow>=9.0.0
